security/manager/ssl/tests/mochitest/mixedcontent/test_dynDelayedUnsecureXHR.html

<!DOCTYPE HTML>
<html>
<head>
  <title>unsecure XHR test</title>
  <script type="text/javascript" src="/MochiKit/Base.js"></script>
  <script type="text/javascript" src="/MochiKit/DOM.js"></script>
  <script type="text/javascript" src="/MochiKit/Style.js"></script>
  <script type="text/javascript" src="/MochiKit/Signal.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mixedContentTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script class="testbody" type="text/javascript">
  /* import-globals-from mixedContentTest.js */
  "use strict";

  hasMixedActiveContent = true;

  function runTest()
  {
    isSecurityState("secure");
    window.setTimeout(() => {
      try {
        let req = new XMLHttpRequest();
        req.open("GET", "http://example.com/tests/security/manager/ssl/tests/mochitest/mixedcontent/alloworigin.sjs", false);
        req.send(null);

        // Change should be immediate, the request was sent synchronously
        isSecurityState("broken", "security broken after insecure XHR");
      } catch (ex) {
        ok(false, ex);
      }

      finish();
    }, 0);
  }

  function afterNavigationTest()
  {
    isSecurityState("secure", "security full after navigation");
    finish();
  }

  </script>
</head>

<body>
</body>
</html>
